Build common interface for Postgres and SQL sequence generator
